Add TourGuide details endpoint and related schema changes
Added GET /api/tourguide/{id} to fetch a tour guide with cities and tour packages. Introduced query/handler/DTOs for this feature. Updated repository to support eager loading of related data. Changed TourGuide.CancellationRate to decimal. Enhanced City entity with CenterCoordinates. Updated AutoMapper profiles. Replaced initial migration to add new tables and relationships for cities, guide packages, and requests. Updated EF model snapshot and configs accordingly.

+++ Fayora.Domain/Entities/TourGuideModule/TourGuide.cs
@@ -24,7 +24,7 @@ public class TourGuide : AuditableEntity<Guid>
     public bool IsOnline { get; private set; }
     public int CompletedToursCount { get; private set; }
     public bool IsSuperGuide { get; private set; }
-    public int CancellationRate { get; private set; } // Percentage of tours cancelled by the guide
+    public decimal CancellationRate { get; private set; } // Percentage of tours cancelled by the guide
     public GeoPoint LastLocation { get; private set; } = new GeoPoint(0, 0);
     public DateTimeOffset? LastLocationUpdate { get; private set; }
     public TransportInfo? TransportInfo { get; private set; }
